requirements.txt
fastapi[standard]
uvicorn
pydantic
mysql-connector-python
python-dotenv